Add files via upload
This work project aims to find methods of identifying the most distant proteins and most diverse subsets of proteins
from large protein databases in a scalable and efficient way using a dataset of protein embeddings from SwissProt,
data mining techniques and metaheuristics. The project implements a brute force calculation of the MaxSum Euclidean
distance for each embedding as well as two variations of traditional tabu searches, an alternative tabu search and a
memetic tabu search.

The project contains the code for the algorithms, validation, testing, data processing, data cleaning, data analysis and
other supplementary code.

This was created for my dissertation project.